refactor: validator 역할 분리
